game over panel shows all tricks now

+++ main/java/de/jskat/data/SkatGameData.java
@@ -723,11 +723,11 @@ public class SkatGameData {
 
 		// TODO what happens if two or more players have the same points?
 		// FIXME this is rule logic --> remove it from data object!!!
-		if (playerPoints.get(Player.FORE_HAND) > playerPoints
-				.get(Player.MIDDLE_HAND)) {
+		if (playerPoints.get(Player.FORE_HAND).intValue() > playerPoints.get(
+				Player.MIDDLE_HAND).intValue()) {
 
-			if (playerPoints.get(Player.FORE_HAND) > playerPoints
-					.get(Player.HIND_HAND)) {
+			if (playerPoints.get(Player.FORE_HAND).intValue() > playerPoints
+					.get(Player.HIND_HAND).intValue()) {
 				ramschLoser = Player.FORE_HAND;
 			} else {
 				ramschLoser = Player.HIND_HAND;
@@ -735,8 +735,8 @@ public class SkatGameData {
 
 		} else {
 
-			if (playerPoints.get(Player.MIDDLE_HAND) > playerPoints
-					.get(Player.HIND_HAND)) {
+			if (playerPoints.get(Player.MIDDLE_HAND).intValue() > playerPoints
+					.get(Player.HIND_HAND).intValue()) {
 				ramschLoser = Player.MIDDLE_HAND;
 			} else {
 				ramschLoser = Player.HIND_HAND;
